+++ src/main/java/com/example/sales/controller/ProductController.java
@@ -35,13 +35,13 @@ public class ProductController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<String> updateProduct(@RequestHeader("Authorization") String authorization, @PathVariable Long id, @RequestBody ProductUpdateRequestDTO request) throws Exception {
+    public ResponseEntity<String> updateProduct(@RequestHeader("Authorization") String authorization, @PathVariable Integer id, @RequestBody ProductUpdateRequestDTO request) throws Exception {
         productService.updateProduct(authorization, id, request);
         return ResponseEntity.ok("The product was successfully updated!");
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<HttpStatus> deleteProductById(@RequestHeader("Authorization") String authorization, @PathVariable Long id) {
+    public ResponseEntity<HttpStatus> deleteProductById(@RequestHeader("Authorization") String authorization, @PathVariable Integer id) {
         productService.deleteProductById(authorization, id);
         return ResponseEntity.ok().build();
     }
